Rename and move ExecSummary to support non-batch executors
