+++ src/coprocessor/dag/batch/executors/selection_executor.rs
@@ -7,7 +7,7 @@ use tipb::expression::Expr;
 use tipb::expression::FieldType;
 
 use super::super::interface::*;
-use crate::coprocessor::dag::batch::statistics::ExecSummaryCollectorDisabled;
+use crate::coprocessor::dag::exec_summary::ExecSummaryCollectorDisabled;
 use crate::coprocessor::dag::expr::{EvalConfig, EvalContext};
 use crate::coprocessor::dag::rpn_expr::{RpnExpression, RpnExpressionBuilder};
 use crate::coprocessor::{Error, Result};
@@ -126,10 +126,10 @@ impl<C: ExecSummaryCollector, Src: BatchExecutor> BatchExecutor for BatchSelecti
 
     #[inline]
     fn next_batch(&mut self, scan_rows: usize) -> BatchExecuteResult {
-        let timer = self.summary_collector.on_start_batch();
+        let timer = self.summary_collector.on_start_iterate();
         let result = self.handle_next_batch(scan_rows);
         self.summary_collector
-            .on_finish_batch(timer, result.data.rows_len());
+            .on_finish_iterate(timer, result.data.rows_len());
         result
     }
