+++ src/coprocessor/dag/mod.rs
@@ -26,6 +26,7 @@ pub mod aggr_fn;
 pub mod batch;
 pub mod batch_handler;
 pub mod builder;
+pub mod exec_summary;
 pub mod executor;
 pub mod expr;
 pub mod handler;
